Update install.md for Py 3.9 dev suggested dependencies
Found issue with scroll mouse scaling in plots for Py 3.8/3.9 when using wxpython 4.1.1 and matplotlib 3.5 or 3.6.  Unrealistic step and button values in the MPL event.  And always 'one direction' regardless of roll direction. Found that Py 3.9 with wxpython 4.2.0 and matplotlib 3.6.2 had proper behaviour.  Tending to think that we just won't support those inbetween versions as the solution to this.